Speed up gopls startup
Go runtime is installed in /go which is word writable so there is no
need for a dedicated user. This removes the need to usermod at runtime
which was very time consuming. Not having to manipulate the user at
runtime allows to remove the entrypoint script entirely.

With this setup the language server is up and running in milliseconds
rather than seconds.

I based this image on official golang alpine image to avoid as much
duplication as possible